provisioner: Add InstallGPU step
The GPU install behavior implemented here is mostly a pure
reimplementation of the shell script that we have in
install_gpu.sh. One of the small differences between this implementation
and install_gpu.sh is that the nvidia installer container is
parameterizable. Some implementation notes are below.

We replace our usage of the unix.Mount function with calls to the
mount(8) command. This is because mount(8) automatically propagates
mount options from the parent mount to the child mount. unix.Mount does
not do this, and implementing that behavior is non-trivial because there
is no good way to find the mount options associated with an existing
mount without parsing /proc/self/mountinfo. We can avoid parsing
/proc/self/mountinfo ourselves by using the mount(8) command.

We also use the Docker command line client to interact with Docker
instead of using the native Golang Docker client. Using the command line
client is simpler for now since it's easier to stub out in tests and it
doesn't change any of the compatibility concerns that COS Customizer
currently has with respect to Docker client/server compatibility. If we
think that the native Golang client will simplify even more, we can
switch to that in the future.